requirements.txt
absl-py==2.1.0
altair==5.3.0
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
exif==1.6.0
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.65.1
h5py==3.11.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib
mdurl==0.1.2
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python-headless
opt-einsum==3.3.0
pandas==2.2.2
pillow==10.4.0
plum-py==0.8.7
protobuf==4.25.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.1
pyparsing==3.1.2
pytz==2024.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
smmap==5.0.1
streamlit==1.37.0
tenacity==8.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0

termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.14.1